requirements.txt
flask
requests
chatterbox-tts
soundfile
torch
torchvision
torchaudio
